Update PointerInputProcessor to use HitPathTracker.
There is one sneaky behavioral change that comes along
with this: PointerInputChanges are no longer dispatched
one at a time across the hierarchy, but instead are
dispatched as a group for each PointerInputNode, but one
at a time.

So if you have 2 changes (y and z) across
PointerInputNodes A and B, where B is a child of Node:

Previosly: y would be dispatched to A and then B (and back
up, down, up, and finally down) and then z would.

Now:  y will be dispatched to A, followed by z, and then y
would be dispatched to B, followed by z, and so on.

In the next CL, y and z will be dispatched as a list to
A, and then B, and so on.  That will be a large CL due to
the interface needing to change for each PointerInputNode
to accept a list of PointerInputChanges instead of a single
one.

Bug: 124525282
Test: All existing tests pass!

Change-Id: I5c8268eadc4dfe42110fedacec3643694b42b543