Defer to container animations
If a FrameLayout containing a Fragment sets animateLayoutChanges=true,
and that Fragment has set custom animations or a transition, there is a
conflict in the Fragment animation. This conflict keeps a Fragment that
is replaced (without being added to the back stack from) from being
destroyed or detached.

This change disables the Fragment animations (and transitions) if
animations on its container have already been enabled. Therefore,
Fragments that are replaced can be properly destroyed.

Test: Added Test and reran old test
Fixes: 116257087
Change-Id: I2cd1f657839b443d00f7ded12d03334f75e9d660