Don't label headings in GridLayoutManager and StaggeredGridLayoutManager
aosp/949357 didn't actually merge the changes I wanted. I think it had to
 do with abandoning a previous cl I based this off of.

The Movies app is facing an issue where one of their items that spans
the screen is not a heading but is labeled as such by GridLayoutManager.

It's problematic that this happens with any view that spans the width
of the screen, so let's get rid of it and allow devs to label their own
headings like they do with linear RVs

The managers should be setting the heading parameter false initially
anyway since that api is deprecated.

Bug: b/117202844
Test: Passed RecyclerViewAccessibilityTest and
RecyclerViewAccessibilityLifecycleTest

Change-Id: I5a16887681616ab3af2259949b3b58f96485dbe6