Make sure getCaptureStage was called after init()
(1) Remove the check of the CaptureStages in ImageCapture constructor
since the check is already appled in issueTakePicture().
(2) Cleanup the code for PROCESSOR_TYPE_NONE in PreviewExtender since
it is duplicated with the onRepeating callback..
(3) Add test to verify the getCaptureStage() has not been invoked
before init().
(4) Add test for event callbacks.
(5) Add test for PreviewExtender getCaptureStages()

Bug: 134477673
Test: ./gradlew bOS, and manually run the test.
Change-Id: I3d66bba99ee0c5145a3047e3951ebe6a8983a823