Fix pop when calling setPrimary before replace
When popping a FragmentTransaction, the operations
you apply are undone in reverse ordering. This means
that when you call setPrimaryNavigationFragment()
before replace(), the replace() is undone first, causing
the current Fragment to be removed from the
FragmentManager before the code for undoing
setPrimaryNavigationFragment runs.

We now ensure that performPrimaryNavigationFragmentChanged
is only called if the Fragment is still attached.

Test: new replacePrimaryNavAfterSetPrimary test passes
Fixes: 134673465
Change-Id: I1a126c265c2350ed4e7d437fb68df29376ff5fcf