Call onVideoSizeChanged when current media item is changed
MediaController returns the VideoSize value of the previous media
item if the current media item does not have a video
(e.g. music file). This CL fixes this by making MediaPlayer retrieve
VideoSize when current media item is changed and call
onVideoSizeChanged.

Also, made PlayerWrapper call onVideoSizeChanged when it is
attached to VideoView.

Bug: 136039839
Test: ./gradlew :media2:integration-tests:testapp:installDebug on API
28, 27 devices
./gradlew :media2:media2-widget:connectedCheck on API 28, 27 devices
Change-Id: I617fc8e27ab3554dd078b3036548fc9239eedf92