Populate setScrollDeltaX/Y for RecyclerView
getScrollDeltaX/Y are populated by view for TYPE_VIEW_SCROLLED
accessibility events. The events are sent in onScrollChanged,
where the difference in old and the new scroll positions are
parameters and are used to populate the events. RecyclerView
passes getScrollX() and getScrollY(), as both the old and the
new parameters. getScrollX/Y() for RecyclerView are always 0.
This should be fixed.

Test: added a log below the fix and verified that hresult and vresult
make sense.
Bug: 131267612

Change-Id: I105553a1b7c34b11e1227d28ab6be2ccae4c42e6