Use density param instead of DensityReceiver for Shape
The usage of the previous signature proved to be confusing and hard to read. Here I am moving density into a parameter for Shapes and CornerSizes.
Plus I migrated CornerSize from the typealias to the interface and made all the implementations to be data classes instead of the lambdas. The usage (corners creation) code will not change. This will allow us to compare corner based Shape instances via equals calls and skip recomposition when the object is updated but represents the same values.

Bug: 129278276
Test: tests are passing, added new tests for equals calls.
Change-Id: Ife75cd8ba3fc7542dc4fb4ca5f84ad9accd66b94