Fix crash + simplify biometric/fingerprint fragment usage
The new fragment constructor for BiometricPrompt doesn't correctly
handle the fingerprint dialog shown for API < 28, resulting in a crash
when attempting to detach the dialog with the client activity's fragment
manager rather than the one for the fragment it is currently attached to.
This commit fixes the issue by using getFragmentManager() to get the
fragment manager for the appropriate host, rather than relying on
getActivity().getSupportFragmentManager().

As it turns out, this same trick works in BiometricFragment and
simplifies the usage of that fragment by not requiring BiometricPrompt
to pass in the appropriate FragmentManager on configuration changes. So
this commit makes that change as well.

Test: On a Pixel 2XL, for both API < 28 and API >= 28:
1. Launch the biometric demo app
2. Select the option to host in a fragment
3. Authenticate with various settings. Verify this no longer crashes
4. Back out of the app and verify this no longer crashes either

Change-Id: I3230282bcede3092a9030ac3458a13d9bf708f5e