Compose UI Modifiers
Taking a cue from CoroutineContext, a Modifier is a linear chain of
specialized elements that provide additional decoration/context for
a UI component. Modifiers serve two design purposes for Compose UI:

The first is to address the limitation we have in composable
functions where it is impractical to statically determine whether
a call to another composable function will emit zero, one, or many
layout nodes. This affects the API design of composables intended to
wrap a single child.

We could choose to tie into Kotlin's contracts system for this or
otherwise lean on the Compose compiler for additional analysis, but this
brings along significant unavoidable complexity for new developers along
with new considerations for the tooling ecosystem, included but not
limited to the editing experience, code review tools and API
compatibility tracking for libraries.

We'd prefer not to create more new concepts at that low of a level
unless the reward for that tradeoff is overwhelming. It's not clear that
the calling syntax enabled by such a change would reach that bar.

Which leads to the second concern: the API ergonomics of single-child
composables. The existing API surface of composables such as Padding
impose a down and to the right march of indentation in source code that
impacts readability. Android developers are used to expressing
decorations such as padding, backgrounds or LayoutParams values as
properties on the affected XML element itself rather than as wrappers.
At the same time we want to retain the composability and extensibility
of the system and not end up with a repeat of android.widget.View, where
a fixed framework superclass becomes host to dozens of cross-cutting but
use case-specific properties.

At a glance, Modifiers as implemented in this proof of concept seem to
address these issues. This change implements modification of measurement
and layout, but the idea could be extended to also support drawing,
hit testing, input handling, semantics, arbitrary LayoutParams-style
arguments to parent layouts and more. The concatenation style also
supports partial compositions of Modifiers to be fed through many layers
of the hierarchy, permitting style-like usages and overlay/override
patterns.

Test: Material Rally sample; existing layout tests
Change-Id: I23f5986675bd8f356d2312f353f80086c69356a0