Fixes some layout rounding issues related to alignment
Align now uses floats for calculations and only rounds at the end
MainAxisAlignment now uses floats for calculating gaps

Previously when trying to align to the end / bottom of a container with
an odd width/height, we would incorrectly first center the alignment,
which caused a 1px rounding error, causing the last 1px of the child to
not be visible. Instead we should use floats until the last rounding
operation, to avoid multiple rounding issues.

For MainAxisAlignment, gaps were calculated using integers, so splitting
a 5px gap into two gaps would incorrectly result in two gaps of 3 px -
this means that every position would be incorrect, instead of just the
positions where we need to round.

Also removes 1f delta from layout assertions and fixes correspondingly
failing tests - we need to be pixel correct here to test alignment, and any
other tests that flake depending on DPI either points to a fault in the
layout model, or incorrect test state.

Bug: b/131411280
Test: AlignTest, uncommented parts of AppBarTest
Change-Id: I6adae3de06108a8291bb931999fec9679f37c107