Add optional Bundle to FragmentTransaction Class APIs
When adding or replacing a Fragment with a
Class<? extends Fragment> (added in b/126124987), allow
developers to pass in an optional Bundle of arguments.

In FragmentTransaction itself, the methods take
a @Nullable Bundle, ensuring that developers are
aware that they are expected to pass arguments through
that method if they want them set.

In the Kotlin extensions, the Bundle is the last
parameter, with a default value of null.

Test: updated tests
Fixes: 137684600
Change-Id: Ic492556e92031d2116d394b561002f338e5b1824