Separate AnimatedDraggable to Draggable and DragValueController
To make possible to add drags without animations and split
responsibilies better, AnimatedDraggable was split to Draggable and
DragValueContrller, with which now is possible to control any value with
drag, for example Float or AnimatedDraggable or something else.
Also, +anchoredControllerByState was added to use internally
in ui-material so synchronization between states of composable and
acnhors of drag can be done easily.

Bug: 137665160
Test: ./gradlew ui:ui-material:connectedCheck
Change-Id: I29e6b8411112e8c9f1aef66244b7ebc998d9485c