Do not allow the fingerprint dialog to be canceled due to touches outside
It's safer to disallow this, since on some devices the sensor is on the
home button, etc, where it's easier to miss. Since the support library
is supposed to simplify developer's lives, we should just default to
preventing this on all pre-P devices, rather than potentially exposing
an API for developers to toggle based on device type (too complicated).

Fixes: 125447052

Test: Touches outside no longer cancel authentication. Back button cancels
      authentication still.

Change-Id: I6639b2ecdf18f82a242fffba63eeb8ca45d53dc9