Ensure FragmentContainerView only allows views of Fragments
FragmentContainerView should not allow adding of arbitrary child views.
The purpose of FragmentContainerView is to serve Fragments, not to be
used as another viewGroup.

This change uses the getViewFragment() method when a view is being
added to check if the view belongs to a Fragment. If getViewFragment()
returns a fragment, that view can be added to the FragmentContainerView,
but if null is returned, an IllegalStateException is thrown.

Test: Added Test, all others still pass
BUG: 136494650
Change-Id: I14635640468bf2ee715531b3032fd45e987334dd