Don't destroy Fragments without calling initState()
The only place where keepActive was used was
when a Fragment was animating to the destroyed
state and then it was re-added before the
animation completes resulting in a Fragment that
had its Lifecycle destroyed and, due to keepActive,
did not have its Lifecycle/other state reset
via initState(). This can leave the Fragment in
an inconsistent state and breaks the invariant
around destroyed being a terminal state for a
Lifecycle object.

Instead, we can cancel the animation where it
is and proceed from that state (since that's a
known valid state). This also allows us to
remove the keepActive flag entirely, simplifying
the logic.

Test: all tests still pass
Change-Id: I3bd86e8c99936835c905a0241d90daf1941159bd