Set the Dialog to have transparent background and no title
Setting the window inside the Dialog not to have a title, because
it added an extra layer of padding for the Dialog.
Setting the window to have a transparent background in order to
allow this functionality and also to remove the corner radius of
the Dialog.
Added a demo which contains the sample.

Test: Manually
Change-Id: I814e45ab83958fceb6ec1145587943d1c57c7da8