media2-session: Fix flaky MediaSession_KeyEventTest
Test had been randomly failed because setUp() fails to start playback
and being notified via OnCompletionListener. It was because MediaPlayer
object was GCed while playback is running, so the listener wasn't
called as expected.

This CL keeps the MediaPlayer instance as a member variable so it
wouldn't be unexpectedly GCed.

Bug: 139173649
Test: Run following tests 10 times on API 25 and 29
      $ media2/session/version-compat-tests/runtest.sh 4
Change-Id: Ie18bc0c3704ed962f959e925c1aeb72292a1dc08