Remove PagedList.pagedSource
Exposing PagedSource<*, T> defines the PagedList to be presenting the
same data type as the PagedSource loads. Once we support arbitrary
transforms on the loaded pages e.g.:

    map { repositoryItem -> UiItem(repositoryItem) }

that will no longer be true - the types can be distinct. Remove this
API for now, we can offer alternatives as needed.

Partly the motivation here is for implementing refresh, which should
be explicit API instead.

Test: ./gradlew paging:paging-common:test

Change-Id: I0c4a696f2ec9c1bc0429507052b31be943bc008d