Fix concave path clipping on APIs before Q
Previously even that for paths we don't use outline based clipping we still were passing the concave path into android Outline which we causing a crash on pre Q devices.
Now we do it only for concat paths and also I added a runtime check to prevent unsupported combination of concave shape with non-zero elevation.

Test: new test in ClipTest
Change-Id: I86961e973ef51c93ac8adcdbc0b91aa1a8ae838d