Fix for incomplete RV measure.
Given an optimization in "autoMeasure" for
when both measure modes are EXACTLY, there were
situations where if RV was first passed
other measuremetns with other measure modes
and then later passed measurements with
different measure modes where the children of
of the RV would not be remeasured and laid out.

This fixes the issue by tracking the last measure
result and then later comparing it to the RV's
laid out dimensions.  If the last measurement was
EXACTLY and was skipped, then during layout we
check to see if the last measurement dimensions
are the same as our current RV layout dimensions,
if they are, we don't need to check children again.
if they aren't we have to measure and lay out children
again.

Fixes: 138734786
Test: ./gradlew recyclerview:recyclerview:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.recyclerview.widget.RecyclerViewMeasureTest

Change-Id: I617a8f92d8c2587771550f81d4df1a14c1850b5f