Call OnPositioned callbacks when the global position has been changed
There are quite a lot of cases when the children want be aware about not only local position changes(like moving the children within the direct parent), but when the global coordinates has been changed(for example when the children are located in some container which is being scrolled. Popups and text selection handlers are interested in such cases.
Instead of calling the callbacks right after the positioning of each LayoutNode now we postpone all the callbacks triggering to the moment when we finish the position stage for all the LayoutNodes.
As a side effect it will make the coordinates usage safer as previously there were a possibility to receive the callback, call coordinates.getParentCoordinates(), but this coordinates are outdated, they would be recalculated as this LayoutNode wasn't yet positioned.
Now we don't stop on the measured LayoutNode when traverse the tree, but call the callbacks for all the callbacks down the tree to allow reaction on the global coordinates updates. It introduces an additional tree pass, but I added a benchmark for the case when OnPositioned called in the hierarchy with 100 components depth and there is no any noticeable difference between this implementation and the version before.
One more optimization: now I reuse LayoutCoordinates object. It is now the same object every time for every LayoutNode.
Also I added convenience methods to take the global position or the position in the root widget for the LayoutCoordinates.

Bug: 138655984
Bug: 136746628
Test: new tests in OnPositionedTest, new benchmark in OnPositionedBenchmark
Change-Id: I1c75523a05bb45b74d4c08ce323ef76bcd5c5b97