Table decorations rework and optimisation
This changes the way table decorations work by putting them inside
a WithConstraints which ensures that the decorations will only be
composed once the Table has determined its offsets. Thus it removes
a 1 frame delay that previously existed when drawing the decorations.

Finally this also introduces an argument which specifies whether
a decoration will be placed above or below the content of the Table.

Test: Ran TableTest and added TableTest#testTable_withDecoration()
Change-Id: Ic515f5be8142ddb44cbc3a8e25845fd653aff567