Replace WithConstraints with ad-hoc solution for Drawers
WithConstraints is a very powerfull tool, however, for now it suffers
from the performance lose, and for whole-screen-wrappers, such ass
Drawers, this performance lose is causing the whole screen to drop
frames.

With b/141458323 it seems like it now also breaks "relayouting without
recomposition on @Model change" usecase.

This CL introduces Layout primitive (private for Drawer) which replaces
sub composition creating and maintenance with
composition+recomposition first time instead of just composition. As a
result, Drawer works very fast (no benchmarks for now) and scrolling
inside of it works great as well.

Bug: b/141458323
Test: DrawerTests passes
Change-Id: Iebc5057ff05de834c4a9bcab7c6d5c79d58e80f5