Adds ColorPalette interface and surrounding infrastructure
This makes some things such as animating color values easy to implement,
and is also early preparation for future light/dark theme support.

The default backing implementation is a to-be-memoized class containing
observable color properties, so that when values in this are changed we
only recompose composables that directly consume these colors.

Unary plus syntax is temporary, and should be going away soon after
compiler changes.

Bug: b/138714596
Test: ColorPaletteTest / ColorPaletteBenchmark
Change-Id: I63d53f99413d3eb156abb35b70bc06fc6797ed58