Avoid crashes when dismissing FingerprintDialogFragment
Currently, it's possible for the biometric library to crash if the user
presses back to dismiss the fingerprint dialog while the device is
changing configuration (e.g., rotating). This commit adds and makes use
of a new method for dismissing this fragment that avoids crashes by
first checking if it is associated with a fragment manager.

Test: ./gradlew biometric:connectedAndroidTest
Test: On Android O, show prompt, rotate device, press back repeatedly.
Before: App crashes
After: App does not crash

Fixes: 141356362
Change-Id: I52d2f32df41881a073a44315950b608a4a4f6ecd