Fix for RepaintBoundaries not being sized properly after attaching
RepaintBoundaries were drawn without providing size for them. As a result onParamsChange() wasn't called as well and the current opacity, elevation or clip shape wasn't applied.
Now we would add newly attached RepaintBoundary in repaintBoundaryChanges so the resizing would be scheduled. I also added asserts so if both of implementation of RepaintBoundaries will be drawn before providing a size we would throw an exception.

Test: It is hard to test this with just RepaintBoundary as seems like it can draw even with not defined size, but I added tests for Opacity, DrawShadow and Clip
Change-Id: I355beed2b5cf67f9afb1195f641bd8686af52681