Reload sqlite SO files upon linking failure
This CL works around a bug in xerial where it fails to load the SO file
and does not retry again. Instead, we just force load it and try 5 times
to do so. see: https://github.com/xerial/sqlite-jdbc/issues/97 for details.

I've also removed the driver unloading since that bug is fixed in
xerial.

Bug: 128688823
Test: development/room/stress-test-room.sh
Change-Id: Idd780339211994f00b02d372371f80e1dcd4b5df