Move ripple drawing to overlay
Current behavior was to draw the ripples below the content - as a background. But actually according to material design specs we always need ripples to be drawn as an overlay (similar to setting ripple drawable as a foreground of a view).
The background behavior still would be possible with introducing Stack and applying Ripple for the container with the same size below the real content. But it is better to make the most common case to be easier to write.

Test: manually, it would be possible to test it with unit test when we move bitmap pixel assertion logic to the common testing module from ui-framework
Change-Id: Ic94c344fabb4759acc2dda8a5a6424b54b7753cc